+++ src/main/java/com/epam/tc/hw4/pages/DifferentPage.java
@@ -1,6 +1,6 @@
 package com.epam.tc.hw4.pages;
 
-import io.qameta.allure.Step;
+import com.epam.tc.hw3.pages.BasePage;
 import org.assertj.core.api.SoftAssertions;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -36,45 +36,37 @@ public class DifferentPage extends BasePage {
     @FindBy(xpath = "//li[contains(text(), 'Colors:')]")
     private WebElement logForDropDownYellow;
 
-    @Step ("Select element checkbox with name 'Water'")
     public void selectCheckBoxWater() {
         checkBoxWater.click();
     }
 
-    @Step ("Select element checkbox with name 'Wind'")
     public void selectCheckBoxWind() {
         checkBoxWind.click();
     }
 
-    @Step ("Select element radio with name 'Selen'")
     public void selectRadioSelen() {
         radioSelen.click();
     }
 
-    @Step ("Select from dropdown list element with name 'Yellow'")
     public void selectDropDownYellow() {
         dropDownYellow.click();
     }
 
-    @Step ("Verify that element checkbox with name 'Water' is selected and there is log for it")
     public void verifySelectedCheckBoxWater() {
         softAssertions.assertThat(logForCheckBoxWater.isDisplayed()).isTrue();
         softAssertions.assertThat(checkBoxWater.isSelected()).isTrue();
     }
 
-    @Step ("Verify that element checkbox with name 'Wind' is selected and there is log for it")
     public void verifySelectedCheckBoxWind() {
         softAssertions.assertThat(logForCheckBoxWind.isDisplayed()).isTrue();
         softAssertions.assertThat(checkBoxWind.isSelected()).isTrue();
     }
 
-    @Step ("Verify that element radio with name 'Selen' is selected and there is log for it")
     public void verifySelectedRadioSelen() {
         softAssertions.assertThat(logForRadioSelen.isDisplayed()).isTrue();
         softAssertions.assertThat(radioSelen.isSelected()).isTrue();
     }
 
-    @Step ("Verify that element from dropdown list with name 'Yellow' is selected and there is log for it")
     public void verifySelectedDropDownYellow() {
         softAssertions.assertThat(logForDropDownYellow.isDisplayed()).isTrue();
         softAssertions.assertThat(dropDownYellow.isSelected()).isTrue();
